+++ Lab00App-build/Lab00App-code/src/main/java/ch/heigvd/res/lab00/Flute.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package ch.heigvd.res.lab00;
+
+/**
+ *
+ * @author Antoine
+ */
+public class Flute implements IInstrument{
+    public String play(){
+        return "trilili";
+    }
+    public int getSoundVolume(){
+        return 2;
+    }
+    public String getColor(){
+        return "grey";
+    }
+}
