+++ Lab00App-build/Lab00App-code/src/main/java/ch/heigvd/res/lab00/IInstrument.java
@@ -1,0 +1,16 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package ch.heigvd.res.lab00;
+
+/**
+ *
+ * @author Antoine
+ */
+public interface IInstrument{
+    String play();
+    int getSoundVolume();
+    String getColor();
+}
